Block cleaning script
It turns out that the script wasn't being called remoteley.
Although that is now fixed with the use of psexec (regretebly).
which comes with the advantage that it works but the downside that the
cleaning script requires to be in the pc that needs to be cleaned.

A new branch will be created where the fixes will be worked on.