Update GetNet and serial number correction
Now checks if the network device has a DNSDomain since (in my personal case)
it was showing a virtual Hamachi board instead of the correct one.

Maybe a more correct way to go about this is show more boards with less filtering,
or maybe regexing the first entry of the hashtable to expect an IPv4
(But I don't know how thoroughly is this enforced)
[RESEARCH REQUIRED]

Serial number was not showing up since the variable for the output was mistyped.